requirements.txt
openai==v1.3.4
python-dotenv
langchain